requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
SQLAlchemy==2.0.25
pymysql==1.1.0
python-multipart==0.0.6
Jinja2==3.1.2
python-dotenv==1.0.0
python-docx==0.8.11
docx2txt==0.8
numpy<2.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.0
pdfplumber==0.9.0
pdf2image==1.17.0
